ESparkWin takes better care of units
When changing units or scale factors, ESparkWin now ensures
that the stimulus amplitudes are within ±10 V on the wire.
This is always helpful.

Also, when switching V → A, the numbers are reduced 10⁶×,
and when switching A → V, they are scaled up 10⁶×, so that
mV become nA and vice versa. This is probably usually helpful.